+++ pulse2percept/utils/__init__.py
@@ -15,8 +15,8 @@
     three_dim
 
 """
-from .base import (PrettyPrint, FreezeError, Frozen, Data, bijective26_name,
-                   cached, gamma)
+from .base import (PrettyPrint, FreezeError, Frozen, Parametrized, Data,
+                   bijective26_name, cached, gamma)
 from .geometry import (cart2pol, pol2cart, delta_angle)
 from .array import is_strictly_increasing, radial_mask, sample, unique
 from .images import center_image, scale_image, shift_image, trim_image
@@ -44,6 +44,7 @@ __all__ = [
     'Frozen',
     'gamma',
     'is_strictly_increasing',
+    'Parametrized',
     'parse_3d_orient',
     'pol2cart',
     'PrettyPrint',